Extract animation helpers
