+++ src/index.ts
@@ -1,44 +1,9 @@
 import Plugin from '@swup/plugin';
 import { matchPath, isPromise } from 'swup';
-import type { Handler, Path, Visit } from 'swup';
+import type { Handler, Visit } from 'swup';
+import { Animation, compileAnimations, CompiledAnimation, getBestAnimationMatch } from './animations.js';
 
 type RequireKeys<T, K extends keyof T> = Partial<T> & Pick<T, K>;
-
-type Animation = {
-	/** The path pattern to match the current url against. */
-	from: Path;
-	/** The path pattern to match the next url against. */
-	to: Path;
-	/** The function to call when the animation is triggered. */
-	out: (done: () => void, data: AnimationData) => void | Promise<void>;
-	/** The function to call when the animation is triggered. */
-	in: (done: () => void, data: AnimationData) => void | Promise<void>;
-};
-
-type CompiledAnimation = Animation & {
-	/** Match function to check if the `from` pattern matches a given URL */
-	matchesFrom: MatchFunction;
-	/** Match function to check if the `to` pattern matches a given URL */
-	matchesTo: MatchFunction;
-};
-
-type AnimationData = {
-	visit: Visit;
-	direction: 'in' | 'out';
-	from: {
-		url: string;
-		pattern: Path;
-		params: object;
-	};
-	to: {
-		url: string;
-		pattern: Path;
-		params: object;
-	};
-};
-
-type MatchOptions = Parameters<typeof matchPath>[1];
-type MatchFunction = ReturnType<typeof matchPath>;
 
 type Options = {
 	/** The selector for matching the main content area of the page. */
@@ -48,6 +13,10 @@ type Options = {
 };
 
 type InitOptions = RequireKeys<Options, 'animations'>;
+
+export type MatchOptions = Parameters<typeof matchPath>[1];
+
+export type MatchFunction = ReturnType<typeof matchPath>;
 
 export default class SwupJsPlugin extends Plugin {
 	name = 'SwupJsPlugin';
@@ -80,7 +49,7 @@ export default class SwupJsPlugin extends Plugin {
 
 		this.options = { ...this.defaults, ...options };
 		this.options.animations.push(this.defaultAnimation);
-		this.animations = this.compileAnimations();
+		this.animations = compileAnimations(this.options.animations, this.options.matchOptions);
 	}
 
 	mount() {
@@ -88,24 +57,15 @@ export default class SwupJsPlugin extends Plugin {
 		this.replace('animation:out:await', this.awaitOutAnimation, { priority: -1 });
 	}
 
-	// Compile path patterns to match functions and transitions
-	compileAnimations(): CompiledAnimation[] {
-		return this.options.animations.map((animation): CompiledAnimation => {
-			const matchesFrom = matchPath(animation.from, this.options.matchOptions);
-			const matchesTo = matchPath(animation.to, this.options.matchOptions);
-			return { ...animation, matchesFrom, matchesTo };
-		});
-	}
-
 	awaitInAnimation: Handler<'animation:in:await'> = async (visit, { skip }) => {
 		if (skip) return;
-		const animation = this.getBestAnimationMatch(visit);
+		const animation = getBestAnimationMatch(this.animations, visit.from.url, visit.to.url, visit.animation.name);
 		await this.createAnimationPromise(animation, visit, 'in');
 	};
 
 	awaitOutAnimation: Handler<'animation:out:await'> = async (visit, { skip }) => {
 		if (skip) return;
-		const animation = this.getBestAnimationMatch(visit);
+		const animation = getBestAnimationMatch(this.animations, visit.from.url, visit.to.url, visit.animation.name);
 		await this.createAnimationPromise(animation, visit, 'out');
 	};
 
@@ -147,49 +107,5 @@ export default class SwupJsPlugin extends Plugin {
 				result.then(resolve);
 			}
 		});
-	}
-
-	getBestAnimationMatch(visit: Visit): CompiledAnimation | null {
-		let topRating = 0;
-
-		const animation: CompiledAnimation | null = this.animations.reduceRight(
-			(bestMatch, animation) => {
-				const rating = this.rateAnimation(visit, animation);
-				if (rating >= topRating) {
-					topRating = rating;
-					return animation;
-				} else {
-					return bestMatch;
-				}
-			},
-			null as CompiledAnimation | null
-		);
-
-		return animation;
-	}
-
-	rateAnimation(visit: Visit, animation: CompiledAnimation): number {
-		const from = visit.from.url;
-		const to = visit.to.url!;
-		const name = visit.animation.name;
-
-		let rating = 0;
-
-		// check if route patterns match
-		const fromMatched = animation.matchesFrom(from);
-		const toMatched = animation.matchesTo(to);
-		if (fromMatched) {
-			rating += 1;
-		}
-		if (toMatched) {
-			rating += 1;
-		}
-
-		// beat all others if custom name fits
-		if (fromMatched && animation.to === name) {
-			rating += 2;
-		}
-
-		return rating;
 	}
 }
